Smart warehouse/agv_simulator/requirements.txt
paho-mqtt
numpy  
networkx